+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * 【poi】  ExcelReader增加getWriter、getOrCreateCell方法
 * 【core】NetUtil增加isInRange方法（感谢@【成都】小邓）
 * 【core】新增BeanPath
+* 【core】CollUtil新增reverse、reverseNew方法
 
 ### Bug修复
 * 【core】修复IdcardUtil中身份证15转18位年的问题（Issue#IHT1Q@Gitee）